Update with Stream refactor comments
