+++ models/src/main/scala/dto/game/properties/MasteryDto.scala
