+++ models/src/main/scala/dto/game/properties/RuneDto.scala
