+++ models/src/main/scala/dto/match/PlayerDto.scala
@@ -1,0 +1,4 @@
+package dto.`match`
+
+case class PlayerDto(currentPlatformId: Option[String], summonerName: Option[String], matchHistoryUri: Option[String], platformId: Option[String],
+                     currentAccountId: Option[String], profileIcon: Option[Int], summonerId: Option[String], accountId: Option[String])
