+++ queries-scala/README.md
@@ -1,0 +1,6 @@
+# Scala Queries
+
+## Purpose
+This module is use to create some queries to the data using Scala collections.
+
+<!--TODO: Explain module-->
