+++ queries-spark/README.md
@@ -1,0 +1,7 @@
+# Spark Queries
+
+## Purpose
+This module is use to create some queries to the data using Spark framework, DataSet's and
+DataFrame's.
+
+<!--TODO: Explain module-->
