--- src/main/resources/application.conf
@@ -1,1 +1,0 @@
-akka.http.client.parsing.ignore-illegal-header-for = [date]
