--- src/main/resources/credentials.properties.template
@@ -1,15 +1,0 @@
-# This is a template of the configuration of the project
-# First change the extension from .properties.template to .properties
-# Enable the properties you want and add more if you need it
-
-# Set your private API key to access the Riot API
-apiKey=PersonalAPIKey
-
-# This is the header of the API key
-riotToken=X-Riot-Token
-
-# Data encoding
-encoding=UTF-8
-
-# Set the path where your dataset will be written
-outputPath=src/outputData/
