--- src/main/scala/usefull/MaxRequestException.scala
@@ -1,3 +1,0 @@
-package usefull
-
-case class MaxRequestException(timeToWait: Int) extends RuntimeException
