+++ utils/src/main/scala/utils/StreamComponents.scala
@@ -1,0 +1,31 @@
+package utils
+
+import akka.NotUsed
+import akka.actor.ActorSystem
+import akka.http.scaladsl.Http
+import akka.http.scaladsl.model.headers.RawHeader
+import akka.http.scaladsl.model.{HttpRequest, HttpResponse, Uri}
+import akka.stream.scaladsl.{Flow, Source}
+import com.typesafe.scalalogging.Logger
+
+import scala.util.{Random, Try}
+
+object StreamComponents {
+  private val logger = Logger("Stream Components")
+
+  //TODO improve the header election, not random.
+  def createRequests(uris: List[String], headers: List[RawHeader]): Source[(HttpRequest, Uri), NotUsed] =
+    Source(uris map (uri => HttpRequest(uri = uri)
+      .withHeaders(getRandomHeader(headers)) // Refactor header usage
+      -> Uri(uri)))
+
+  def sendRequest(host: String)
+                 (implicit actorSystem: ActorSystem): Flow[(HttpRequest, Uri), (Try[HttpResponse], Uri), Http.HostConnectionPool] =
+    Http().cachedHostConnectionPoolHttps[Uri](host = host)
+
+  def createRequestRetry(uris: List[Uri], headers: List[RawHeader]): Source[(HttpRequest, Uri), NotUsed] =
+    Source(uris.map(uri => HttpRequest(uri = uri).withHeaders(getRandomHeader(headers)) -> uri))
+
+  private def getRandomHeader(headers: List[RawHeader]) = headers(Random.nextInt(headers.length))
+
+}
