+++ utils/src/main/scala/utils/TimeMeasure.scala
@@ -1,0 +1,9 @@
+package utils
+
+object TimeMeasure {
+  def apply[T](f: => T): (T, Long) = {
+    val start = System.nanoTime()
+    val res = f
+    (res, System.nanoTime() - start)
+  }
+}
