+++ utils/src/main/scala/utils/Uris.scala
@@ -1,0 +1,8 @@
+package utils
+
+object Uris {
+  val riotChallengerUri: String = "/lol/league/v4/challengerleagues/by-queue/RANKED_SOLO_5x5"
+  val riotSummonerUri: String = "/lol/summoner/v4/summoners/"
+  val riotMatchlistUri: String = "/lol/match/v4/matchlists/by-account/"
+  val riotMatchUri: String = "/lol/match/v4/matches/"
+}
